+++ akka-typed-tests/src/test/scala/docs/akka/typed/IntroSpec.scala
@@ -11,8 +11,6 @@ import scala.concurrent.Future
 import scala.concurrent.duration._
 import scala.concurrent.Await
 //#imports
-import akka.testkit.AkkaSpec
-import akka.typed.TypedSpec
 
 object IntroSpec {
 
@@ -50,7 +48,10 @@ object IntroSpec {
     //#chatroom-protocol
     //#chatroom-behavior
 
-    def chatRoom(sessions: List[ActorRef[SessionEvent]] = List.empty): Behavior[Command] =
+    val behavior: Behavior[Command] =
+      chatRoom(List.empty)
+
+    private def chatRoom(sessions: List[ActorRef[SessionEvent]]): Behavior[Command] =
       Stateful[Command] { (ctx, msg) ⇒
         msg match {
           case GetSession(screenName, client) ⇒
@@ -66,28 +67,6 @@ object IntroSpec {
         }
       }
     //#chatroom-behavior
-
-    //#chatroom-actor
-    def mutableChatRoom(): Behavior[Command] =
-      Deferred[Command] { _ =>
-        var sessions: List[ActorRef[SessionEvent]] = List.empty
-        Stateful[Command] { (ctx, msg) ⇒
-          msg match {
-            case GetSession(screenName, client) ⇒
-              val wrapper = ctx.spawnAdapter {
-                p: PostMessage ⇒ PostSessionMessage(screenName, p.message)
-              }
-              client ! SessionGranted(wrapper)
-              sessions = client :: sessions
-              Same
-            case PostSessionMessage(screenName, message) ⇒
-              val mp = MessagePosted(screenName, message)
-              sessions foreach (_ ! mp)
-              Same
-          }
-        }
-      }
-    //#chatroom-actor
   }
   //#chatroom-actor
 
@@ -136,8 +115,8 @@ class IntroSpec extends TypedSpec {
 
     //#chatroom-main
     val main: Behavior[akka.NotUsed] =
-      Deferred { ctx =>
-        val chatRoom = ctx.spawn(ChatRoom.chatRoom(), "chatroom")
+      Deferred { ctx ⇒
+        val chatRoom = ctx.spawn(ChatRoom.behavior, "chatroom")
         val gabblerRef = ctx.spawn(gabbler, "gabbler")
         ctx.watch(gabblerRef)
         chatRoom ! GetSession("ol’ Gabbler", gabblerRef)
@@ -150,8 +129,7 @@ class IntroSpec extends TypedSpec {
               Stopped
             case _ ⇒
               Unhandled
-          }
-        )
+          })
       }
 
     val system = ActorSystem("ChatRoomDemo", main)
